+++ src/main/java/com/gui/Mogura.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.gui;
 
 import java.awt.*;
 import java.awt.event.*;
